+++ vite.config.js
@@ -1,6 +1,7 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 import { quasar } from "@quasar/vite-plugin";
+import { fileURLToPath, URL } from 'node:url'
 
 const host = process.env.TAURI_DEV_HOST;
 
@@ -39,4 +40,20 @@ export default defineConfig(async () => ({
       }
     }
   },
+
+  // Test configuration
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    setupFiles: ['./tests/setup.js'],
+    deps: {
+    }
+  },
+
+  resolve: {
+    alias: {
+      '@': fileURLToPath(new URL('./src', import.meta.url))
+    }
+  }
+
 }));